command/v2/rename_command_test.go
package v2_test

import (
	"code.cloudfoundry.org/cli/actor/actionerror"
	"code.cloudfoundry.org/cli/command/commandfakes"
	"code.cloudfoundry.org/cli/command/flag"
	. "code.cloudfoundry.org/cli/command/v2"
	"code.cloudfoundry.org/cli/command/v2/v2fakes"
	"code.cloudfoundry.org/cli/util/ui"
	. "github.com/onsi/ginkgo"
	. "github.com/onsi/gomega"
	. "github.com/onsi/gomega/gbytes"
)

var _ = Describe("Rename Command", func() {
	var (
		cmd             RenameCommand
		testUI          *ui.UI
		fakeConfig      *commandfakes.FakeConfig
		fakeSharedActor *commandfakes.FakeSharedActor
		fakeActor       *v2fakes.FakeRenameActor
		binaryName      string
		executeErr      error

		oldName string
		newName string
	)

	BeforeEach(func() {
		testUI = ui.NewTestUI(nil, NewBuffer(), NewBuffer())
		fakeConfig = new(commandfakes.FakeConfig)
		fakeSharedActor = new(commandfakes.FakeSharedActor)
		fakeActor = new(v2fakes.FakeRenameActor)

		oldName = "some-old-app-name"
		newName = "some-new-app-name"

		cmd = RenameCommand{
			RequiredArgs: flag.AppRenameArgs{
				OldAppName: oldName,
				NewAppName: newName,
			},

			UI:          testUI,
			Config:      fakeConfig,
			SharedActor: fakeSharedActor,
			Actor:       fakeActor,
		}

		binaryName = "faceman"
		fakeConfig.BinaryNameReturns(binaryName)
	})

	JustBeforeEach(func() {
		executeErr = cmd.Execute(nil)
	})

	Context("when checking target fails", func() {
		BeforeEach(func() {
			fakeSharedActor.CheckTargetReturns(actionerror.NotLoggedInError{BinaryName: binaryName})
		})

		It("returns an error if the check fails", func() {
			Expect(executeErr).To(MatchError(actionerror.NotLoggedInError{BinaryName: "faceman"}))

			Expect(fakeSharedActor.CheckTargetCallCount()).To(Equal(1))
			checkTargetedOrg, checkTargetedSpace := fakeSharedActor.CheckTargetArgsForCall(0)
			Expect(checkTargetedOrg).To(BeTrue())
			Expect(checkTargetedSpace).To(BeTrue())
		})
	})

	// Context("when the user is logged in, and org and space are targeted", func() {
	// 	BeforeEach(func() {
	// 		fakeConfig.TargetedOrganizationReturns(configv3.Organization{Name: "some-org"})
	// 		fakeConfig.TargetedSpaceReturns(configv3.Space{
	// 			GUID: "some-space-guid",
	// 			Name: "some-space"})
	// 		fakeConfig.CurrentUserReturns(
	// 			configv3.User{Name: "some-user"},
	// 			nil)
	// 	})
	// })
})
